Frontend: sketch out desugaring of DOT binary expressions
Dot expressions are a special case because their right hand side
argument must be an identifier. Identifiers by themselves don't map to
any type of runtime value, so it is not just a function call.

It could be turned to function call that passes a string, but I don't
like that design because it kinda weakens the type system.