Tests: refactor and implement array size test
Implements Assert::array_of_size. Extracts a function Assert::of_type that takes a Type::Value and doesn't print `Success` at the end, it's a helper function. But if we change the fact that the test utility must print that (Tester could do it instead), we could start using it as a regular test function without further considerations.