TestUtils and Execute: refactor for better translation unit organization
Now test_utils functions will not do the process of creating nodes and such, it is the test's responsability. Now they takes values instead of strings and no Environment is required.
To compensate, the expression node evaluator has been moved to execute.xpp, since a similar deal is happening there already. Now main.cpp has access to it, but the tests need this file as well.