Implement Tarjan's SCC decomposition algorithm (#107)
* Implement Tarjan's SCC decomposition algorithm
* Slightly refactor tests to use a single Tester
* Add tests for Tarjan's algorithm

To separate declarations into rec-blocks, we need to find the strongly
connected components of the graph formed by declarations referencing
each other. (Because recursion can be seen as a cycle in the call graph)

Once we get that, we need to find a topological sort of that
condensation graph. Luckily, Tarjan's algorithm gives us the SCCs already
topologically sorted, so there is no need for additional computation.