fix a bug
We were using hidden vars on local declarations, this turns out to be
wrong, and I am not sure why I thought it was right, tbh.

I do know why it's wrong, though: when the type of a local variable is
unified with a newer type var (e.g. the result of a call expression),
the newer var may end up pointing to the hidden type var, causing
generalization when it sohuld not.