implement match_identifiers
Since we have an identifier that needs to be added to a scope, along
with an optional typehint, I figured we could use a declaration to
write a bit less code.

Declarations are quite bulky at compile time, but I don't think it
matters that much. If it does, we can fix it later.

I feel like I missed something, but can't remember what...