Parser: (bugfix) ensure binop rhs parses successfully
We weren't checking that the rhs of a binary expression was
an error. This would make us crash at a later stage on
expressions like:

   x + + y;

This was fixed with a single use of the TRY macro.

With the addition of another cosmetic change, this means
that we no longer reference Writer<T>::m_result or
Writer<T>::ok() in the entire parser.

In other words, error handling in the parser is now done
exclusively through the TRY/REQUIRE family of macros.