`TRY` macro for the parser and other improvements (#311)
* Parser: replace CHECK_AND_RETURN with the new TRY_WITH macro

TRY_WITH does the same thing as CHECK_AND_RETURN, but it
also evaluates to the m_result of its argument.

This means we can use it inside expressions and avoid
writing .m_result too often.

* Parser: replace CHECK_AND_EXTRACT with new TRY macro

* Parser: take advantage of coercion in Writer's constructors

Writer<T*> can be implicitly converted to Writer<U*> if T is
a subclass of U, so we don't need to write
make_writer<CST::CST*>(e) when we want to return a (for
example) CST::Identifier* as a CST::CST*.

Instead, we can just use make_writer(e) and let it build a
Writer<CST::Identifier*> which will then be implicitly
converted to a Writer<CST::CST*>.

* Parser: take advantage of an equivalence

consider the snippet:
   return make_writer(TRY(x));

it will do the following:
- if x is an error, it returns the error (TRY).
- if not, it unwraps its value (TRY), re-wraps it
  (make_writer), and returns it.

In other words:
- if x is an error, it returns x.
- if x not, it returns x.

Or more simply:
- it returns x.

therefore, it is equivalent to:

   return x;

* Utils: add ErrorReport::add_sub_error method

* Parser: embrace REQUIRE and REQUIRE_WITH macros, and ErrorReport->Writer<T> conversion

* Parser: make error reports less nested

In particular, things like:

   fn f() {
     if (false) {
       if (false) {
		 <syntax-error>;
   }

Would produce errors like:

   - Failed to parse function declaration
   -- Failed to parse block statement
   --- Failed to parse statement
   ---- Failed to parse if statement or expression
   ----- Failed to parse statement
   ------ Failed to parse block statement
   ------- Failed to parse statement
   -------- Failed to parse if statement or expression
   --------- Failed to parse statement
   ---------- Failed to parse block statement
   ----------- Failed to parse statement
   ------------ Failed to parse expression
   ------------- At line:col : <the actual error>

Now, we get a nicer and simpler error:

   - Failed to parse function declaration
   -- At line:col : <the actual error>

* Parser: (bugfix) ensure binop rhs parses successfully

We weren't checking that the rhs of a binary expression was
an error. This would make us crash at a later stage on
expressions like:

   x + + y;

This was fixed with a single use of the TRY macro.

With the addition of another cosmetic change, this means
that we no longer reference Writer<T>::m_result or
Writer<T>::ok() in the entire parser.

In other words, error handling in the parser is now done
exclusively through the TRY/REQUIRE family of macros.