Parser: [WIP] Create tools to show context for syntax errors
When a syntax error (or any kind of error, really) is found,
we'd like to show the place in the file where it happened.

These functions can be as inefficient as we like because,
given that we exit right after, they run at most one time
per run of the executable.