Remove token array (#316)
* Frontend: use a std::vector to store tokens

In the past, we would lex and parse concurrently. This meant
that we needed a container that would maintain reference
stability as tokens were added. std::vector::push_back does
not have this property, so we created ChunkedArray.

We now fully tokenize files before parsing, so this
requirement has been lifted. For simplicity's sake, we're
using a vector.

* Frontend: remove TokenArray type alias

* Frontend: Make tokens 30% smaller

Due to pointer alignment, tokens were full with padding.
Reordering the members removes padding and speeds up the
lexer+parser by about 12%.