Typechecker: Implement the typing environment more simply
In HM systems with the value restriction, free type variables of non-values get
"bound to the environment". This is not an explicit operation. Rather, it means
that those variables occur in the type of a variable that is bound to the typing
environment.

"The typing environment" is just a fancy name for "a list of variables that are
available in the current scope, along with their types". Since we already store
types in declarations, this means we just need some data structure that allows
us to store declarations and remove them according to Jasper's scoping rules.

Of course, already have such a data structure: it's called SymbolTable. So we
can use that instead of the ad-hoc CompileTimeEnvironment.