Don't deep-copy types when generalizing
We used to deep-copy and replace the generalized variables
with fresh variables, in case they were unified and a
polytype was silently mutated.

Turns out that this case is actually impossible:

We only generalize variables when we are done processing a
declaration and the variable doesn't appear free in the
environment.

Since the variable doesn't appear free in the environment,
there is no way we can even access it later on, so there is
no way it can be unified.