Typechecker: Implement the typing environment in a simpler way (#318)
* Typechecker: Implement the typing environment more simply

In HM systems with the value restriction, free type variables of non-values get
"bound to the environment". This is not an explicit operation. Rather, it means
that those variables occur in the type of a variable that is bound to the typing
environment.

"The typing environment" is just a fancy name for "a list of variables that are
available in the current scope, along with their types". Since we already store
types in declarations, this means we just need some data structure that allows
us to store declarations and remove them according to Jasper's scoping rules.

Of course, already have such a data structure: it's called SymbolTable. So we
can use that instead of the ad-hoc CompileTimeEnvironment.

* Typechecker: rename typecheck() to infer() and make it static

The only public function in this module is typecheck_program.
Everything else should be static.

* Don't deep-copy types when generalizing

We used to deep-copy and replace the generalized variables
with fresh variables, in case they were unified and a
polytype was silently mutated.

Turns out that this case is actually impossible:

We only generalize variables when we are done processing a
declaration and the variable doesn't appear free in the
environment.

Since the variable doesn't appear free in the environment,
there is no way we can even access it later on, so there is
no way it can be unified.

* Refactor the main typecheck_program loop

* Tooling: print each negative test that runs

* Interpreter: fix a bug in Value::print

Looks like I broke printing after reordering some declarations.

All the C++ pointer types were getting coerced to bool, so
all Jasper reference types were getting printed as Booleans.