Add a bytecode interpreter (#319)
* Eval: simplify how CallExpression is evaluated

This makes the handling of Function and NativeFunction
almost the same, which makes it easier to compile to
bytecode.

* Bytecode: create the bytecode interpreter

Bytecode is a flat and compact representation of code
that can be interpreted with little overhead. To get some
bytecode to run we first need to compile some Jasper
code into that form. This can be done in a single pass
over the AST.

This commit adds an AST to bytecode compiler and a
simple bytecode interpreter. The current bytecode
doesn't support all of Jasper right now.

Each AST::FunctionLiteral will now store a pointer to its
bytecode representation. Initially NULL, it gets filled in
after the first call to that function. If some unsupported
is needed, compilation fails and we fall back on the old
tree-walking interpreter.