Metacheck again (#321)
* Typechecker: change metacheck to not use unification

Turns out we can do metatype inference in a fully top-down
manner if we first do a shallow inference pass and order the
declarations carefully after that for a deep pass.

The new implementation is inspired by bidirectional
type-checking: it uses two functions named check and infer
and -- although we haven't added any yet -- it leaves open
the possibility of special adding special cases in check() for
better accuracy and error messages.

The main idea is that the metatype of most expressions can
be inferred just from the syntactic category (i.e. the
ASTTag of the node). The only expressions that can't be
inferred this way are Identifiers (that inherit their
metatype from their declaration) and access expressions
(that depend on the metatype of their subexpression).

To ensure identifiers can be inferred, we must visit them
after we infer a metatype for their declaration, so we visit
declarations in dependency order.

If there ever is a cycle of declarations that can't be
inferred shallowly, it's either really weird code, or just
broken. In those cases we just error out. Even in the cases
where the code is not broken, we don't much care about
supporting this kind of code for now.

Example 1 - wrong

    a := b;
    b := c;
    c := a;

Example 2 - wrong

    a := b.x;
    b := a.y;

Example 3 - wrong

    a := c(b).x
    b := d(a).y

Example 4 - ok but really weird

    a := c(fn() => b).x
    b := d(fn() => a).y

* Typechecker: take advantage of the completeness of metatype inference

This lets us change some error messages in ct_eval to
straight-up assertions.

* Typechecker: blast meta_unifier into orbit