Typechecker: remove sus code in TypeSystemCore::new_term()
I inlined all the calls and it the end it just creates a
typefunc with strength=None and points its union-find node
at the given tf, which accomplishes absolutely nothing.

The original idea is that unification would constraint the
typefunc var to have the right number of arguments, but we
are steering away from the idea of typefunc variables now,
so it's better to remove it than to "fix" it.