always unwrap gc_ptr after allocation
This required slight adjustments in some places. The crux
was that any value that is not on the stack or on the global
scope when a GC allocation happens may be collected, so we
need to push/store things as soon as possible, or delay
allocation until as late as possible.

In particular, recursive calls to eval() will almost
certainly perform gc allocations, so we need to make sure we
don't hold any potentially-collectable objects when calling
eval().